requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
starlette==0.27.0
SQLAlchemy==2.0.17
pydantic==1.10.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.24.1
psycopg2-binary==2.9.3

